test: move BasePersistenceTest & BaseServiceTest
