+++ src/test/java/com/meoguri/linkocean/domain/notification/service/NotificationServiceImplTest.java
@@ -15,7 +15,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Slice;
 
 import com.meoguri.linkocean.domain.notification.entity.Notification;
-import com.meoguri.linkocean.test.support.service.BaseServiceTest;
+import com.meoguri.linkocean.test.support.domain.service.BaseServiceTest;
 
 class NotificationServiceImplTest extends BaseServiceTest {
