+++ src/test/java/com/meoguri/linkocean/domain/profile/command/persistence/CustomProfileRepositoryImplTest.java
@@ -9,7 +9,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import com.meoguri.linkocean.domain.profile.command.entity.Profile;
-import com.meoguri.linkocean.test.support.persistence.BasePersistenceTest;
+import com.meoguri.linkocean.test.support.domain.persistence.BasePersistenceTest;
 
 class CustomProfileRepositoryImplTest extends BasePersistenceTest {
