+++ src/test/java/com/meoguri/linkocean/domain/profile/command/persistence/FindProfileByIdQueryTest.java
@@ -9,7 +9,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Import;
 
 import com.meoguri.linkocean.domain.profile.command.entity.Profile;
-import com.meoguri.linkocean.test.support.persistence.BasePersistenceTest;
+import com.meoguri.linkocean.test.support.domain.persistence.BasePersistenceTest;
 
 @Import(FindProfileByIdQuery.class)
 class FindProfileByIdQueryTest extends BasePersistenceTest {
