+++ src/test/java/com/meoguri/linkocean/domain/profile/command/service/FavoriteServiceImplTest.java
@@ -12,7 +12,7 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import com.meoguri.linkocean.domain.bookmark.service.dto.GetDetailedBookmarkResult;
-import com.meoguri.linkocean.test.support.service.BaseServiceTest;
+import com.meoguri.linkocean.test.support.domain.service.BaseServiceTest;
 
 //TODO 서비스 보강 (예상하는대로 쿼리가 안나감?)
 class FavoriteServiceImplTest extends BaseServiceTest {
