+++ src/test/java/com/meoguri/linkocean/domain/profile/query/service/ProfileQueryServiceImplTest.java
@@ -15,7 +15,7 @@ import org.springframework.data.domain.Slice;
 import com.meoguri.linkocean.domain.profile.query.persistence.dto.ProfileFindCond;
 import com.meoguri.linkocean.domain.profile.query.service.dto.GetDetailedProfileResult;
 import com.meoguri.linkocean.domain.profile.query.service.dto.GetProfilesResult;
-import com.meoguri.linkocean.test.support.service.BaseServiceTest;
+import com.meoguri.linkocean.test.support.domain.service.BaseServiceTest;
 
 class ProfileQueryServiceImplTest extends BaseServiceTest {
