+++ src/test/java/com/meoguri/linkocean/domain/tag/service/TagServiceImplTest.java
@@ -11,7 +11,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import com.meoguri.linkocean.domain.profile.query.service.dto.GetProfileTagsResult;
-import com.meoguri.linkocean.test.support.service.BaseServiceTest;
+import com.meoguri.linkocean.test.support.domain.service.BaseServiceTest;
 
 class TagServiceImplTest extends BaseServiceTest {
